Urls Provider, NamedStringProvider
Added the `NamedStringProvider`. The `NamedStringProvider` is a returned a
function used to return named string from a string collection in the service
configuration. For example the configuration could specify a key named `urls`,
which contains a key/value pair of URLs used by the application. A
`NamedServiceProvider` can be registered with `urls` as it's collection named,
and the provider function can be used to retrieve the values by key.

The `Paths` service provider was refactored to extend `NamedStringProvider`, and
the `Urls` provider was added as another common `NamedCollectionProvider`.